Move SharedMap to managed/map.ts
This allows other Shared objects to be present in their own files.
