+++ assembly/index.ts
@@ -2,7 +2,7 @@ export * from "./distributed";
 export * from "./distributed/bindings";
 export * from "./error";
 export * from "./error/bindings";
-export * from "./managed";
+export * from "./managed/map";
 export * from "./managed/held";
 export * from "./managed/maybe";
 export * from "./managed/yieldable";
