+++ assembly/managed/map.ts
